[Refactor] move vm.py under runtime and adt to runtime.container.py (#4855)
