+++ python/tvm/container.py
@@ -17,7 +17,7 @@
 """Container data structures used in TVM DSL."""
 import tvm._ffi
 
-from tvm.runtime import Object, ObjectTypes
+from tvm.runtime import Object
 from tvm.runtime.container import getitem_helper
 from tvm.runtime import _ffi_node_api
 from . import _api_internal
@@ -104,56 +104,3 @@ class LoweredFunc(Object):
     MixedFunc = 0
     HostFunc = 1
     DeviceFunc = 2
-
-
-@tvm._ffi.register_object("vm.ADT")
-class ADT(Object):
-    """Algebatic data type(ADT) object.
-
-    Parameters
-    ----------
-    tag : int
-        The tag of ADT.
-
-    fields : list[Object] or tuple[Object]
-        The source tuple.
-    """
-    def __init__(self, tag, fields):
-        for f in fields:
-            assert isinstance(f, ObjectTypes), "Expect object or " \
-            "tvm NDArray type, but received : {0}".format(type(f))
-        self.__init_handle_by_constructor__(_ADT, tag, *fields)
-
-    @property
-    def tag(self):
-        return _GetADTTag(self)
-
-    def __getitem__(self, idx):
-        return getitem_helper(
-            self, _GetADTFields, len(self), idx)
-
-    def __len__(self):
-        return _GetADTSize(self)
-
-
-def tuple_object(fields=None):
-    """Create a ADT object from source tuple.
-
-    Parameters
-    ----------
-    fields : list[Object] or tuple[Object]
-        The source tuple.
-
-    Returns
-    -------
-    ret : ADT
-        The created object.
-    """
-    fields = fields if fields else []
-    for f in fields:
-        assert isinstance(f, ObjectTypes), "Expect object or tvm " \
-        "NDArray type, but received : {0}".format(type(f))
-    return _Tuple(*fields)
-
-
-tvm._ffi._init_api("tvm.container")
